+++ DearPyGui/src/core/AppItems/basic/mvDragMulti.cpp
@@ -238,11 +238,11 @@ namespace Marvel {
                 auto value = *_value;
                 if (_alias.empty())
                     GContext->callbackRegistry->submitCallback([=]() {
-                    GContext->callbackRegistry->addCallback(getCallback(false), _uuid, ToPyFloatList(value.data(), value.size()), _user_data);
+                    GContext->callbackRegistry->addCallback(getCallback(false), _uuid, ToPyFloatList(value.data(), (int)value.size()), _user_data);
                         });
                 else
                     GContext->callbackRegistry->submitCallback([=]() {
-                    GContext->callbackRegistry->addCallback(getCallback(false), _alias, ToPyFloatList(value.data(), value.size()), _user_data);
+                    GContext->callbackRegistry->addCallback(getCallback(false), _alias, ToPyFloatList(value.data(), (int)value.size()), _user_data);
                         });
             }
         }
@@ -520,11 +520,11 @@ namespace Marvel {
                 auto value = *_value;
                 if (_alias.empty())
                     GContext->callbackRegistry->submitCallback([=]() {
-                    GContext->callbackRegistry->addCallback(getCallback(false), _uuid, ToPyIntList(value.data(), value.size()), _user_data);
+                    GContext->callbackRegistry->addCallback(getCallback(false), _uuid, ToPyIntList(value.data(), (int)value.size()), _user_data);
                         });
                 else
                     GContext->callbackRegistry->submitCallback([=]() {
-                    GContext->callbackRegistry->addCallback(getCallback(false), _alias, ToPyIntList(value.data(), value.size()), _user_data);
+                    GContext->callbackRegistry->addCallback(getCallback(false), _alias, ToPyIntList(value.data(), (int)value.size()), _user_data);
                         });
             }
         }
